requirements.txt
numpy==1.19.5
pandas==1.3.3
scikit-learn==1.0.2
scikit-image==0.19.3
protobuf==3.17.3
tensorflow-cpu==2.4.4
